Impose a limit on sigchecks density.
We allow at most 1 sigcheck for every 45 bytes of scriptSig length, plus
a free allowance of 2 sigchecks. Note that the total length of an input
is equal to 40 + len(scriptSig) + len(varint), where len(varint) is
typically 1 or 2 bytes. Thus, the maximal density of sigchecks in a
transaction (with empty scriptSigs) is 1 per 20.5 bytes.

Only three old transactions in BCH chain have violated this limit;
the limit is designed to allow the two most-dense standard cases:
- spending bare multisig 1-of-3 in legacy mode (3 sigchecks and 72 byte scriptsig)
- spending p2sh-multisig 1-of-15 in legacy mode (15 sigchecks and 588 byte scriptsig)